replace old figure file with content of figureManager file
